+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/formatters/base_formatter.rb
@@ -1,0 +1,13 @@
+module RubyLsp
+  module Ree
+    class BaseFormatter
+      def self.call(source, uri)
+        new.call(source, uri)
+      end
+
+      def call(source, uri)
+        raise 'abstract method'
+      end
+    end
+  end
+end