+++ ruby-lsp-ree/spec/ruby_lsp_ree_helper.rb
@@ -1,7 +1,9 @@
 module RubyLspReeHelper
-  def index_fn(server, name, package = nil)
+  def index_fn(server, name, package = nil, uri = nil)
     location = RubyIndexer::Location.new(0, 0, 0, 0)
-    file_uri = if package
+    file_uri = if uri
+      uri
+    elsif package
       URI("file:///#{package}/package/#{package}/#{name}.rb")
     else
       URI("file:///#{name}.rb")
@@ -69,5 +71,42 @@ module RubyLspReeHelper
         position: position
       }
     )
+  end
+
+  def sample_package_dir
+    @sample_package_dir ||= File.expand_path(
+      File.join(__dir__, 'sample_package')
+    )
+  end
+
+  def sample_package_locales_dir
+    File.join(sample_package_dir, 'package', 'sample_package', 'locales')
+  end
+
+  def sample_file_uri
+    file_name = 'my_file'
+    package_name = 'sample_package'
+    URI("file://#{sample_package_dir}/package/#{package_name}/#{file_name}.rb")
+  end
+
+  def ruby_document(source)
+    RubyLsp::RubyDocument.new(
+      source: source, 
+      version: 1, 
+      uri: URI.parse(''), 
+      global_state: RubyLsp::GlobalState.new
+    )
+  end
+
+  def store_locales_cache
+    {
+      en: File.read(sample_package_locales_dir + '/en.yml'),
+      ru: File.read(sample_package_locales_dir + '/ru.yml')
+    }
+  end
+
+  def restore_locales_cache(cache)
+    File.write(sample_package_locales_dir + '/en.yml', cache[:en])
+    File.write(sample_package_locales_dir + '/ru.yml', cache[:ru])
   end
 end